Unify processor logic
